fix: [#1880] Replace setTimeout with queueMicrotask where appropriate
This commit replaces several instances of `setTimeout(..., 0)` with
`queueMicrotask` to improve performance and align more closely with
browser behavior for microtask scheduling.

The main changes are:

- In `MutationObserverListener`, mutation observer callbacks are now scheduled with `queueMicrotask`, as specified by the DOM standard.
- In `Fetch`, the creation of a `ReadableStream` from a buffer now uses `queueMicrotask` to make the operation asynchronous more efficiently.
- In `BrowserWindow`, `postMessage` now uses `queueMicrotask` to dispatch the `message` event.

I've ignored the `setTimeout` in
`@packages/happy-dom/src/window/BrowserWindow.ts` where it's being used
to simulate smooth scrolling, since in that case the macrotask
scheduling is necessary to accurately reflect DOM behavior.